+++ versiondb/tsrocksdb/store.go
@@ -240,6 +240,13 @@ func (s Store) Flush() error {
 	)
 }
 
+// Close does not flush the memtable; call Flush first if durability is needed.
+func (s Store) Close() error {
+	s.cfHandle.Destroy()
+	s.db.Close()
+	return nil
+}
+
 // FixData fixes wrong data written in versiondb due to rocksdb upgrade, the operation is idempotent.
 // see: https://github.com/crypto-org-chain/cronos-store/issues/1683
 // call this before `SetSkipVersionZero(true)`.